+++ src/plans/augmented_lagrangian_plan.jl
@@ -10,6 +10,7 @@ _doc_AL_Cost_long = raw"""
 \Bigr)
 ```
 """
+
 @doc """
     AugmentedLagrangianCost{CO,R,T}
 
@@ -30,7 +31,7 @@ number type used and ``T`` the vector type.
 
     AugmentedLagrangianCost(co, ρ, μ, λ)
 """
-mutable struct AugmentedLagrangianCost{CO,R,T}
+mutable struct AugmentedLagrangianCost{CO,R,T} <: AbstractConstrainedFunctor{T}
     co::CO
     ρ::R
     μ::T
@@ -40,14 +41,8 @@ function set_manopt_parameter!(alc::AugmentedLagrangianCost, ::Val{:ρ}, ρ)
     alc.ρ = ρ
     return alc
 end
-function set_manopt_parameter!(alc::AugmentedLagrangianCost, ::Val{:μ}, μ)
-    alc.μ = μ
-    return alc
-end
-function set_manopt_parameter!(alc::AugmentedLagrangianCost, ::Val{:λ}, λ)
-    alc.λ = λ
-    return alc
-end
+get_manopt_parameter(alc::AugmentedLagrangianCost, ::Val{:ρ}) = alc.ρ
+
 function (L::AugmentedLagrangianCost)(M::AbstractManifold, p)
     gp = get_inequality_constraint(M, L.co, p, :)
     hp = get_equality_constraint(M, L.co, p, :)
@@ -60,12 +55,8 @@ function (L::AugmentedLagrangianCost)(M::AbstractManifold, p)
     return c + (L.ρ / 2) * d
 end
 
-_doc_AL_Grad = raw"""
-
-"""
-
-@doc """
-    AugmentedLagrangianGrad{CO,R,T}
+@doc raw"""
+    AugmentedLagrangianGrad{CO,R,T} <: AbstractConstrainedFunctor{T}
 
 Stores the parameters ``ρ ∈ ℝ``, ``μ ∈ ℝ^m``, ``λ ∈ ℝ^n``
 of the augmented Lagrangian associated to the [`ConstrainedManifoldObjective`](@ref) `co`.
@@ -90,7 +81,7 @@ number type used and ``T`` the vector type.
     AugmentedLagrangianGrad(co, ρ, μ, λ)
 
 """
-mutable struct AugmentedLagrangianGrad{CO,R,T}
+mutable struct AugmentedLagrangianGrad{CO,R,T} <: AbstractConstrainedFunctor{T}
     co::CO
     ρ::R
     μ::T
@@ -100,20 +91,11 @@ function (LG::AugmentedLagrangianGrad)(M::AbstractManifold, p)
     X = zero_vector(M, p)
     return LG(M, X, p)
 end
-
 function set_manopt_parameter!(alg::AugmentedLagrangianGrad, ::Val{:ρ}, ρ)
     alg.ρ = ρ
     return alg
 end
-function set_manopt_parameter!(alg::AugmentedLagrangianGrad, ::Val{:μ}, μ)
-    alg.μ = μ
-    return alg
-end
-function set_manopt_parameter!(alg::AugmentedLagrangianGrad, ::Val{:λ}, λ)
-    alg.λ = λ
-    return alg
-end
-
+get_manopt_parameter(alg::AugmentedLagrangianGrad, ::Val{:ρ}) = alg.ρ
 # default, that is especially when the `grad_g` and `grad_h` are functions.
 function (LG::AugmentedLagrangianGrad)(
     M::AbstractManifold, X, p, range=NestedPowerRepresentation()