+++ src/solvers/convex_bundle_method.jl
@@ -64,8 +64,8 @@ Most of the following keyword arguments set default values for the fields mentio
 mutable struct ConvexBundleMethodState{
     P,
     T,
-    Pr,
-    St,
+    Pr<:Union{F,AbstractManoptProblem} where {F},
+    St<:AbstractManoptSolverState,
     R,
     A<:AbstractVector{<:R},
     B<:AbstractVector{Tuple{<:P,<:T}},
@@ -77,7 +77,7 @@ mutable struct ConvexBundleMethodState{
     TS<:Stepsize,
     TSC<:StoppingCriterion,
     VT<:AbstractVectorTransportMethod,
-} <: AbstractManoptSolverState where {R<:Real,P,T,I<:Int,Pr,St}
+} <: AbstractManoptSolverState where {R<:Real,P,T,I<:Int,Pr}
     atol_λ::R
     atol_errors::R
     bundle::B
@@ -122,7 +122,7 @@ mutable struct ConvexBundleMethodState{
         X::T=zero_vector(M, p),
         vector_transport_method::VT=default_vector_transport_method(M, typeof(p)),
         sub_problem::Pr=convex_bundle_method_subsolver,
-        sub_state::St=AllocatingEvaluation(),
+        sub_state::Union{AbstractEvaluationType,AbstractManoptSolverState}=AllocatingEvaluation(),
         k_size=nothing,# deprecated
         p_estimate=nothing,# deprecated
         ϱ=nothing,# deprecated
@@ -132,7 +132,6 @@ mutable struct ConvexBundleMethodState{
         P,
         T,
         Pr,
-        St,
         I,
         TM<:AbstractManifold,
         TR<:AbstractRetractionMethod,
@@ -141,6 +140,7 @@ mutable struct ConvexBundleMethodState{
         VT<:AbstractVectorTransportMethod,
         R<:Real,
     }
+        sub_state_storage = maybe_wrap_evaluation_type(sub_state)
         bundle = Vector{Tuple{P,T}}()
         g = zero_vector(M, p)
         last_stepsize = one(R)
@@ -155,7 +155,7 @@ mutable struct ConvexBundleMethodState{
             P,
             T,
             Pr,
-            St,
+            typeof(sub_state_storage),
             R,
             typeof(linearization_errors),
             typeof(bundle),
@@ -192,7 +192,7 @@ mutable struct ConvexBundleMethodState{
             ξ,
             λ,
             sub_problem,
-            sub_state,
+            sub_state_storage,
             ϱ,# deprecated
         )
     end
@@ -347,7 +347,7 @@ function convex_bundle_method!(
     ),
     vector_transport_method::VTransp=default_vector_transport_method(M, typeof(p)),
     sub_problem=convex_bundle_method_subsolver,
-    sub_state=evaluation,
+    sub_state::Union{AbstractEvaluationType,AbstractManoptSolverState}=evaluation,
     k_size=nothing,# deprecated
     p_estimate=nothing,# deprecated
     ϱ=nothing,# deprecated
@@ -356,6 +356,7 @@ function convex_bundle_method!(
     sgo = ManifoldSubgradientObjective(f, ∂f!!; evaluation=evaluation)
     dsgo = decorate_objective!(M, sgo; kwargs...)
     mp = DefaultManoptProblem(M, dsgo)
+    sub_state_storage = maybe_wrap_evaluation_type(sub_state)
     bms = ConvexBundleMethodState(
         M,
         p;
@@ -372,7 +373,7 @@ function convex_bundle_method!(
         stopping_criterion=stopping_criterion,
         vector_transport_method=vector_transport_method,
         sub_problem=sub_problem,
-        sub_state=sub_state,
+        sub_state=sub_state_storage,
         k_size=k_size,# deprecated
         p_estimate=p_estimate,# deprecated
         ϱ=ϱ,# deprecated
@@ -476,7 +477,7 @@ end
 # Dispatching on different types of subsolvers
 # (a) closed form allocating
 function _convex_bundle_subsolver!(
-    M, bms::ConvexBundleMethodState{P,T,F,AllocatingEvaluation}
+    M, bms::ConvexBundleMethodState{P,T,F,ClosedFormSubSolverState{AllocatingEvaluation}}
 ) where {P,T,F}
     bms.λ = bms.sub_problem(
         M, bms.p_last_serious, bms.linearization_errors, bms.transported_subgradients
@@ -485,7 +486,7 @@ function _convex_bundle_subsolver!(
 end
 # (b) closed form in-place
 function _convex_bundle_subsolver!(
-    M, bms::ConvexBundleMethodState{P,T,F,InplaceEvaluation}
+    M, bms::ConvexBundleMethodState{P,T,F,ClosedFormSubSolverState{InplaceEvaluation}}
 ) where {P,T,F}
     bms.sub_problem(
         M, bms.λ, bms.p_last_serious, bms.linearization_errors, bms.transported_subgradients