+++ test/plans/test_state.jl
@@ -89,6 +89,10 @@ struct NoIterateState <: AbstractManoptSolverState end
         @test d2.state.X == ones(3)
         @test get_stopping_criterion(d2) === s2.stop
     end
+    @testset "Closed Form State" begin
+        @test Manopt.ClosedFormSubSolverState() isa
+            Manopt.ClosedFormSubSolverState{AllocatingEvaluation}
+    end
     @testset "Generic Objective and State solver returns" begin
         f(M, p) = 1
         o = ManifoldCostObjective(f)